requirements.txt
# Dependencies

arxiv>=2.1.0
google-generativeai>=0.8.0
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.9.0
httpx>=0.27.0
tqdm>=4.66.0
moviepy>=2.2.0
Pillow>=11.0.0
